[libc++][NFC] Remove some dead code in string (#94893)
It looks like the last references got removed in c747bd0e2339.
It removed a __zero() function, which was probably created at
some point in the ancient past to optimize copying the string
representation. The __zero() function got simplified to an
assignment as part of making string constexpr, rendering this
code unnecessary.

NOKEYCHECK=True
GitOrigin-RevId: cce1feb7b1caf6bd5fab311a787481a3f6a9dcdf