[libc++] Add default copy ctor to "__chrono/exception.h" (#95338)
After PR#90394, "__chrono/exception.h" will trigger
"deprecated-copy-with-user-provided-dtor" warning on Windows x64 runtime
testing with ToT Clang. This patch addresses this issue by explicitly
adding those default copy ctors.

It is a bit weird that the same warning will not happen when testing on
Linux x64 under the same condition, despite the warning flag was enabled
(with `-Wdeprecated-copy -Wdeprecated-copy-dtor`). It might be a bug.

NOKEYCHECK=True
GitOrigin-RevId: 8b9dce333f71bc21b3534e89a41e1ea8672aa063