[libcxx] [ci] Simplify the mingw x86_64 configs (#95609)
Since switching the Windows CI environment over to GitHub Actions
runners, the mingw tests run in a setup where the default compiler
binary is a mingw-targeting compiler, so we don't need to specify a
custom executable name.

For the i686 tests, we still specify a custom compiler name, in order to
target i686 instead of x86_64.

NOKEYCHECK=True
GitOrigin-RevId: 67285feffd6708e6db36c11faf95eeab449e797a