[libc++] Make sure the test for compare-and-wait bug doesn't hang forever (#97907)
This patch adds an explicit timeout mechanism in the compare-and-wait
test for std::atomic, ensuring that it doesn't run forever when the bug
is present. This is not an issue when we run inside the CI because we
specify a timeout manually, but it can be a problem when running
locally, for example.

NOKEYCHECK=True
GitOrigin-RevId: 0d2621641ca54315dafe8f1aecda6b68db6aa5c8