[libc++] Add clang-20 to failing tests on Windows (#100119)
After we switched to LLVM version 20, some libc++ tests started failing
on Windows. This patch adds the clang-20 condition to XFAIL to fix the
issue. The way that these tests are excluded from Windows are fragile
and need to be updated every time we bump the LLVM version.

NOKEYCHECK=True
GitOrigin-RevId: 1c3a99c7397ea630949ab65a87fa6dd97609e4b3