Fix fs.op.copy_file/copy_file_large.pass.cpp on z/OS (#100382)
This PR is required to fix copy_file_large.pass.cpp which is failing on
z/OS in all ASCII variations. The problem is that a destination file is
opened in binary mode and auto-conversion does not happen when this lit
is compiled with -fzos-le-char-mode=ascii.

In addition opening a destination file will match a text mode of
`fopen()` of a source file.

NOKEYCHECK=True
GitOrigin-RevId: fd17064dfa08c252c12b71167bc588f57e6a082c