[libc++][memory_resource] Implements LWG3683. (#100775)
The polymorphic_allocator was added in C++17.
This issue was filed in 2022 so well after C++20. This issue adds an
operator==.

Starting with C++20 this adds a compiler generated operator!=. To have
the same behaviour in C++17 and C++20 (and later) a manual operator!= is
defined in C++17.

Implements
- LWG3683 operator== for polymorphic_allocator cannot deduce template
argument in common cases

NOKEYCHECK=True
GitOrigin-RevId: 2711618214b22bbc170edd0672f0b5fc1cb8acf4