[libc++][spaceship] Removes unneeded relational operators. (#100441)
This is a followup of https://github.com/llvm/llvm-project/pull/99343.
Since that patch was quite late in the LLVM-19 release cycle some of the
unneeded relational operator were not removed in C++20.

This removes them and gives the change a bit more "baking" time, just in
case there are issues with this change in user code. This change is
intended to be an NFC.

NOKEYCHECK=True
GitOrigin-RevId: 39b6900852e7a1187bd742ba5c1387ca1be58e2c