[libcxx] [test] Mark gdb_pretty_printer_test.sh.cpp as unsupported on Windows (#102891)
In practice, this test hasn't been run by any of the Windows
configurations so far, because it has been excluded by the "UNSUPPORTED:
clang-18, clang-19" line - but it does end up failing if running with
e.g. clang-20 on Windows.

As the test fails on more fundamental issues on Windows, mark it
outright unsupported on this platform; this fixes testing libcxx with a
recent nightly version of Clang from git main.

NOKEYCHECK=True
GitOrigin-RevId: a2acea5c380534430a01843698d4bf46b018110f