[libcxx] Add fallback to standard C when `unistd` is unavailable (#102005)
Summary:
This utility function gets a temp file to use for tests. It either uses
WIN32 or POSIX to create it. Some targets only follow the C standard,
and this test case will fail. This patch simply adds a fallback that
uses the `tmpnam` function from standard C. This function isn't ideal,
but it is good enough for our use-case.

---------

Co-authored-by: Mark de Wever <zar-rpg@xs4all.nl>
NOKEYCHECK=True
GitOrigin-RevId: 3b7b2f2efc8be3077b919bef12669a85886840b5