[libcxx] Disable invalid `__start/__stop` reference on NVPTX (#99381)
Summary:
The logic for this `__is_function_overridden` check requires accessing a
runtime array normally created by the linker. The NVPTX target is an
`__ELF__` target, however it does not support emitting the
`__start/__stop` symbols for C-identifier named sections. This needs to
be disabled explicitly so that the user can compile this with anything.
NOKEYCHECK=True
GitOrigin-RevId: 23617f2dc094e3f960d9e0aedaa87a0866e8dc2d